LPS-66368 move logic to base class
