+++ modules/apps/foundation/portal/portal-upgrade/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Portal Upgrade
 Bundle-SymbolicName: com.liferay.portal.upgrade
-Bundle-Version: 2.3.2
+Bundle-Version: 2.4.0
 Export-Package:\
 	com.liferay.portal.upgrade.registry,\
 	com.liferay.portal.upgrade.release
