+++ modules/apps/foundation/portal/portal-upgrade/src/main/resources/com/liferay/portal/upgrade/release/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.0
+version 1.1.0
